move dispute/kill system workflows out of the unit suite into e2e
