+++ test/unit/DisputeManager.test.ts
@@ -909,79 +909,6 @@ describe("Unit: DisputeManager", function () {
                 h.event.getEventCallCount(peer.index, "onInitiatingDispute")
             ).to.equal(1);
         });
-
-        it("real fault on a settled fork → dispute() applies the fraud proof via multicall, no calldata", async function () {
-            const h = TestSession.getHarness();
-            await h.scenario.preDisputeSetup({ peerCount: 4 });
-            const offender = await h.query.getNextPeerToWrite();
-
-            // the honest disputers detect the invalid transition, store a fraud
-            // proof for the offender, and dispute -> constructDispute bundles it
-            // so dispute() takes the multicall branch (apply proofs + upload)
-            await h.byzantine.submitInvalidStateTransitionBlock(offender.index);
-            // settled fork -> the dispute is final, no auditing calldata posted
-            await h.assert.dispute.initiatedAndCommitedWait({
-                initiatedWithAuditingData: false
-            });
-            await h.dispute.resolveDisputeWait({ forkId: h.activeForkId! });
-
-            // the multicall applied the bundled fraud proof -> offender slashed
-            const witness = h.peers.find((p) => p.index !== offender.index)!;
-            const r = await h.execOnHost(
-                witness,
-                async (sm, args) => {
-                    const slashes = (
-                        await sm.diamondStateMachine.localDiamondContract.getOnChainSlashedParticipants(
-                            sm.channelId
-                        )
-                    ).map((a) => String(a).toLowerCase());
-                    return { offenderSlashed: slashes.includes(args.offender) };
-                },
-                { offender: offender.address.toLowerCase() }
-            );
-
-            expect(r.offenderSlashed).to.equal(true);
-        });
-
-        it("real fault on an unfinalized (pending-join) fork → dispute posted WITH auditing calldata", async function () {
-            const h = TestSession.getHarness();
-            // "calldata-backed" = the pending inbound join leaves the head
-            // not-final-by-everyone, so the dispute's postedAuditingData is true
-            // and dispute() takes the with-calldata upload (vs the settled-fork
-            // test above, which posts without calldata). shorter evidence time
-            // keeps the real resolve under budget
-            await h.scenario.preDisputeSetupCalldataPath({
-                timeConfig: { evidenceTime: 6 }
-            });
-            const offender = await h.query.getNextPeerToWrite();
-
-            await h.byzantine.submitInvalidStateTransitionBlock(offender.index);
-            // the diff from the settled-fork test: this dispute carries calldata
-            await h.assert.dispute.initiatedAndCommitedWait({
-                initiatedWithAuditingData: true
-            });
-            await h.dispute.resolveDisputeWait({
-                forkId: h.activeForkId!,
-                forkSettleTimeoutMs: 20000,
-                syntheticOnChainParticipants: 1
-            });
-
-            const witness = h.peers.find((p) => p.index !== offender.index)!;
-            const r = await h.execOnHost(
-                witness,
-                async (sm, args) => {
-                    const slashes = (
-                        await sm.diamondStateMachine.localDiamondContract.getOnChainSlashedParticipants(
-                            sm.channelId
-                        )
-                    ).map((a) => String(a).toLowerCase());
-                    return { offenderSlashed: slashes.includes(args.offender) };
-                },
-                { offender: offender.address.toLowerCase() }
-            );
-
-            expect(r.offenderSlashed).to.equal(true);
-        });
     });
 
     describe("killDispute", function () {
@@ -1241,85 +1168,6 @@ describe("Unit: DisputeManager", function () {
             const after = await h.query.onChainSlashedParticipants();
             expect(after.length).to.equal(before.length + 1);
             expect(after).to.include(spammer.address.toLowerCase());
-        });
-
-        it("a killed spam dispute re-killed → no second slash, no throw", async function () {
-            const h = TestSession.getHarness();
-            await h.scenario.preDisputeSetup({
-                timeConfig: { evidenceTime: 6 }
-            });
-
-            // a dispute that's internally valid but has no enforcement basis ->
-            // honest peers audit-fail, store a fraud proof, kill it, slash the
-            // spammer (peer 1) on-chain
-            await h.tamper.postTamperedDispute(1, (dispute) => {
-                dispute.input.timeout.participant =
-                    "0x0000000000000000000000000000000000000000";
-                dispute.input.onChainSlashes = [];
-                dispute.input.selfRemoval = false;
-            });
-            await h.event.waitForAllPeers("onDisputeKilled", 1, {
-                mode: "atLeast"
-            });
-
-            // the original detached kill (peer 0's EventHandler) shares peer 0's
-            // signer with the re-kill below, so drain it first - two competing
-            // applyDisputeFraudProofs txs would race on the nonce
-            expect(
-                (await h.quiesceHosts()).map((e) => e.message)
-            ).to.deep.equal([]);
-
-            // re-kill from the stored proof: the spammer is already slashed, so
-            // the on-chain apply is idempotent -> killDispute completes with no
-            // throw and no second slash
-            const r = await h.execOnHost(
-                h.getPeer(0),
-                async (sm) => {
-                    const proofs =
-                        sm.storage.disputeFraudProofs.getDisputeFraudProofs();
-                    if (proofs.length === 0) return { hadProof: false };
-                    const before = (
-                        await sm.diamondStateMachine.localDiamondContract.getOnChainSlashedParticipants(
-                            sm.channelId
-                        )
-                    ).map(String);
-                    let threw = "";
-                    try {
-                        await sm.disputeManager.killDispute(proofs[0].dispute);
-                    } catch (e) {
-                        threw = e instanceof Error ? e.message : String(e);
-                    }
-                    const after = (
-                        await sm.diamondStateMachine.localDiamondContract.getOnChainSlashedParticipants(
-                            sm.channelId
-                        )
-                    ).map(String);
-                    return {
-                        hadProof: true,
-                        threw,
-                        beforeCount: before.length,
-                        afterCount: after.length
-                    };
-                },
-                {},
-                // two on-chain reads + a re-kill's on-chain apply can overrun the
-                // default 2s control-RPC budget under load
-                { timeoutMs: 20000 }
-            );
-
-            // the re-kill is the only in-flight apply -> it must settle clean
-            expect(
-                (await h.quiesceHosts()).map((e) => e.message)
-            ).to.deep.equal([]);
-
-            expect(r.hadProof).to.equal(true);
-            expect(r.threw).to.equal("");
-            expect(r.afterCount).to.equal(r.beforeCount);
-
-            await h.dispute.resolveDisputeWait({
-                forkId: h.activeForkId!,
-                forkSettleTimeoutMs: 15000
-            });
         });
     });
